hello-python/requirements.txt
fastapi[standard]
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi